apps/api/pyproject.toml
[project]
name = "china-outbound-analyzer-api"
version = "0.1.0"
description = "FastAPI backend for the China Outbound Stock AI Analyzer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "akshare>=1.17.87,<2.0.0",
  "alembic>=1.17.1,<2.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "fastapi>=0.135.3,<0.136.0",
  "httpx>=0.28.1,<0.29.0",
  "numpy>=1.26,<2.0.0",
  "orjson>=3.11.4,<4.0.0",
  "pandas>=2.3.3,<3.0.0",
  "psycopg[binary]>=3.2.12,<4.0.0",
  "pydantic-settings>=2.11.0,<3.0.0",
  "sqlalchemy>=2.0.44,<3.0.0",
  "structlog>=25.4.0,<26.0.0",
  "tenacity>=9.1.2,<10.0.0",
  "uvicorn[standard]>=0.38.0,<0.39.0"
]

[dependency-groups]
dev = [
  "pytest>=8.4.2,<9.0.0",
  "pytest-asyncio>=1.2.0,<2.0.0",
  "ruff>=0.14.1,<0.15.0"
]

[project.scripts]
outbound-analyzer = "china_outbound_analyzer.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
ignore = ["E501"]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"migrations/env.py" = ["E402"]
"migrations/versions/*.py" = ["E501", "I001", "UP007", "UP035"]

[tool.hatch.build.targets.wheel]
packages = ["src/china_outbound_analyzer"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
